+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/InAppWebView/InAppWebViewClient.java
@@ -6,12 +6,14 @@ import android.net.http.SslCertificate;
 import android.net.http.SslError;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Message;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.webkit.ClientCertRequest;
 import android.webkit.CookieManager;
 import android.webkit.CookieSyncManager;
 import android.webkit.HttpAuthHandler;
+import android.webkit.RenderProcessGoneDetail;
 import android.webkit.SafeBrowsingResponse;
 import android.webkit.SslErrorHandler;
 import android.webkit.ValueCallback;
@@ -20,15 +22,20 @@ import android.webkit.WebResourceResponse;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
+import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
+import androidx.webkit.WebViewFeature;
 
 import com.pichillilorenzo.flutter_inappwebview.CredentialDatabase.Credential;
 import com.pichillilorenzo.flutter_inappwebview.CredentialDatabase.CredentialDatabase;
 import com.pichillilorenzo.flutter_inappwebview.InAppBrowser.InAppBrowserActivity;
 import com.pichillilorenzo.flutter_inappwebview.JavaScriptBridgeInterface;
+import com.pichillilorenzo.flutter_inappwebview.Shared;
 import com.pichillilorenzo.flutter_inappwebview.Util;
 
 import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -47,7 +54,7 @@ import io.flutter.plugin.common.MethodChannel;
 
 public class InAppWebViewClient extends WebViewClient {
 
-  protected static final String LOG_TAG = "IABWebViewClient";
+  protected static final String LOG_TAG = "IAWebViewClient";
   private FlutterWebView flutterWebView;
   private InAppBrowserActivity inAppBrowserActivity;
   public MethodChannel channel;
@@ -358,7 +365,7 @@ public class InAppWebViewClient extends WebViewClient {
           }
         }
 
-        handler.cancel();
+        InAppWebViewClient.super.onReceivedHttpAuthRequest(view, handler, host, realm);
       }
 
       @Override
@@ -368,7 +375,7 @@ public class InAppWebViewClient extends WebViewClient {
 
       @Override
       public void notImplemented() {
-        handler.cancel();
+        InAppWebViewClient.super.onReceivedHttpAuthRequest(view, handler, host, realm);
       }
     });
   }
@@ -483,7 +490,7 @@ public class InAppWebViewClient extends WebViewClient {
           }
         }
 
-        handler.cancel();
+        InAppWebViewClient.super.onReceivedSslError(view, handler, error);
       }
 
       @Override
@@ -493,7 +500,7 @@ public class InAppWebViewClient extends WebViewClient {
 
       @Override
       public void notImplemented() {
-        handler.cancel();
+        InAppWebViewClient.super.onReceivedSslError(view, handler, error);
       }
     });
   }
@@ -552,7 +559,7 @@ public class InAppWebViewClient extends WebViewClient {
           }
         }
 
-        request.cancel();
+        InAppWebViewClient.super.onReceivedClientCertRequest(view, request);
       }
 
       @Override
@@ -562,20 +569,28 @@ public class InAppWebViewClient extends WebViewClient {
 
       @Override
       public void notImplemented() {
-        request.cancel();
+        InAppWebViewClient.super.onReceivedClientCertRequest(view, request);
       }
     });
   }
 
   @Override
   public void onScaleChanged(WebView view, float oldScale, float newScale) {
+    super.onScaleChanged(view, oldScale, newScale);
     final InAppWebView webView = (InAppWebView) view;
     webView.scale = newScale;
+
+    Map<String, Object> obj = new HashMap<>();
+    if (inAppBrowserActivity != null)
+      obj.put("uuid", inAppBrowserActivity.uuid);
+    obj.put("oldScale", oldScale);
+    obj.put("newScale", newScale);
+    channel.invokeMethod("onScaleChanged", obj);
   }
 
   @RequiresApi(api = Build.VERSION_CODES.O_MR1)
   @Override
-  public void onSafeBrowsingHit(final WebView view, WebResourceRequest request, int threatType, final SafeBrowsingResponse callback) {
+  public void onSafeBrowsingHit(final WebView view, final WebResourceRequest request, final int threatType, final SafeBrowsingResponse callback) {
     Map<String, Object> obj = new HashMap<>();
     if (inAppBrowserActivity != null)
       obj.put("uuid", inAppBrowserActivity.uuid);
@@ -608,7 +623,7 @@ public class InAppWebViewClient extends WebViewClient {
           }
         }
 
-        callback.showInterstitial(true);
+        InAppWebViewClient.super.onSafeBrowsingHit(view, request, threatType, callback);
       }
 
       @Override
@@ -618,7 +633,7 @@ public class InAppWebViewClient extends WebViewClient {
 
       @Override
       public void notImplemented() {
-        callback.showInterstitial(true);
+        InAppWebViewClient.super.onSafeBrowsingHit(view, request, threatType, callback);
       }
     });
   }
@@ -627,6 +642,13 @@ public class InAppWebViewClient extends WebViewClient {
   public WebResourceResponse shouldInterceptRequest(WebView view, final String url) {
 
     final InAppWebView webView = (InAppWebView) view;
+
+    if (webView.options.useShouldInterceptRequest) {
+      WebResourceResponse onShouldInterceptResponse = onShouldInterceptRequest(url);
+      if (onShouldInterceptResponse != null) {
+        return onShouldInterceptResponse;
+      }
+    }
 
     URI uri;
     try {
@@ -696,8 +718,135 @@ public class InAppWebViewClient extends WebViewClient {
   @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
   @Override
   public WebResourceResponse shouldInterceptRequest(WebView view, WebResourceRequest request) {
+    final InAppWebView webView = (InAppWebView) view;
+
     String url = request.getUrl().toString();
+
+    if (webView.options.useShouldInterceptRequest) {
+      WebResourceResponse onShouldInterceptResponse = onShouldInterceptRequest(request);
+      if (onShouldInterceptResponse != null) {
+        return onShouldInterceptResponse;
+      }
+    }
+
     return shouldInterceptRequest(view, url);
+  }
+
+  public WebResourceResponse onShouldInterceptRequest(Object request) {
+    String url = request instanceof String ? (String) request : null;
+    String method = "GET";
+    Map<String, String> headers = null;
+    Boolean hasGesture = false;
+    Boolean isForMainFrame = true;
+    Boolean isRedirect = false;
+
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && request instanceof WebResourceRequest) {
+      WebResourceRequest webResourceRequest = (WebResourceRequest) request;
+      url = webResourceRequest.getUrl().toString();
+      headers = webResourceRequest.getRequestHeaders();
+      hasGesture = webResourceRequest.hasGesture();
+      isForMainFrame = webResourceRequest.isForMainFrame();
+      if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+        isRedirect = webResourceRequest.isRedirect();
+      }
+    }
+
+    final Map<String, Object> obj = new HashMap<>();
+    if (inAppBrowserActivity != null)
+      obj.put("uuid", inAppBrowserActivity.uuid);
+    obj.put("url", url);
+    obj.put("method", method);
+    obj.put("headers", headers);
+    obj.put("isForMainFrame", isForMainFrame);
+    obj.put("hasGesture", hasGesture);
+    obj.put("isRedirect", isRedirect);
+
+    Util.WaitFlutterResult flutterResult;
+    try {
+      flutterResult = Util.invokeMethodAndWait(channel, "shouldInterceptRequest", obj);
+    } catch (InterruptedException e) {
+      e.printStackTrace();
+      Log.e(LOG_TAG, e.getMessage());
+      return null;
+    }
+
+    if (flutterResult.error != null) {
+      Log.e(LOG_TAG, flutterResult.error);
+    }
+    else if (flutterResult.result != null) {
+      Map<String, Object> res = (Map<String, Object>) flutterResult.result;
+      byte[] data = (byte[]) res.get("data");
+      return new WebResourceResponse(res.get("content-type").toString(), res.get("content-encoding").toString(), new ByteArrayInputStream(data));
+    }
+
+    return null;
+  }
+
+  @Override
+  public void onFormResubmission (final WebView view, final Message dontResend, final Message resend) {
+    Map<String, Object> obj = new HashMap<>();
+    if (inAppBrowserActivity != null)
+      obj.put("uuid", inAppBrowserActivity.uuid);
+    obj.put("url", view.getUrl());
+
+    channel.invokeMethod("onFormResubmission", obj, new MethodChannel.Result() {
+
+      @Override
+      public void success(@Nullable Object response) {
+        Map<String, Object> responseMap = (Map<String, Object>) response;
+        Integer action = (Integer) responseMap.get("action");
+        if (action != null) {
+          switch (action) {
+            case 1:
+              resend.sendToTarget();
+              return;
+            case 0:
+            default:
+              dontResend.sendToTarget();
+              return;
+          }
+        }
+
+        InAppWebViewClient.super.onFormResubmission(view, dontResend, resend);
+      }
+
+      @Override
+      public void error(String errorCode, @Nullable String errorMessage, @Nullable Object errorDetails) {
+        Log.d(LOG_TAG, "ERROR: " + errorCode + " " + errorMessage);
+      }
+
+      @Override
+      public void notImplemented() {
+        InAppWebViewClient.super.onFormResubmission(view, dontResend, resend);
+      }
+    });
+  }
+
+  @Override
+  public void onPageCommitVisible(WebView view, String url) {
+    super.onPageCommitVisible(view, url);
+    Map<String, Object> obj = new HashMap<>();
+    if (inAppBrowserActivity != null)
+      obj.put("uuid", inAppBrowserActivity.uuid);
+    obj.put("url", url);
+    channel.invokeMethod("onPageCommitVisible", obj);
+  }
+
+  @RequiresApi(api = Build.VERSION_CODES.O)
+  @Override
+  public boolean onRenderProcessGone(WebView view, RenderProcessGoneDetail detail) {
+    Boolean didCrash = detail.didCrash();
+    Integer rendererPriorityAtExit = detail.rendererPriorityAtExit();
+
+    Map<String, Object> obj = new HashMap<>();
+    if (inAppBrowserActivity != null)
+      obj.put("uuid", inAppBrowserActivity.uuid);
+    obj.put("didCrash", didCrash);
+    obj.put("rendererPriorityAtExit", rendererPriorityAtExit);
+
+    channel.invokeMethod("onRenderProcessGone", obj);
+
+    return super.onRenderProcessGone(view, detail);
   }
 
   @Override