+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/InAppWebViewFileProvider.java
@@ -1,0 +1,9 @@
+package com.pichillilorenzo.flutter_inappwebview;
+
+import androidx.core.content.FileProvider;
+
+public class InAppWebViewFileProvider extends FileProvider {
+
+  // This class intentionally left blank.
+
+}